refactor to avoid component being recreated on every render
