+++ client/app/components/game/PlayerPanel/index.ts
@@ -1,0 +1,1 @@
+export { PlayerPanel as default } from './PlayerPanel'
