+++ client/app/components/game/page/InGame/index.ts
@@ -1,0 +1,1 @@
+export { InGame as default } from './InGame'
