+++ client/app/components/game/page/TeamSelection/TeamSelection.tsx
@@ -1,31 +1,15 @@
 import Panel from '~/components/Panel'
-import { useGameClient, useGameState } from '../GameContext'
-import ListPanel from '~/components/ListPanel'
+import { useGameClient, useGameState } from '../../GameContext'
 import { TeamId } from '~/client/team-id'
 import Button from '~/components/Button'
-import { LuCopy, LuPlus } from 'react-icons/lu'
+import TeamColumn from './TeamColumn'
+import RoomIdLabel from './RoomIdLabel'
 
-interface TeamColumnProps {
-	teamId: TeamId
-	members: string[]
-	onJoin: (teamId: TeamId) => void
-}
-
-const TeamColumn = ({ teamId, members, onJoin }: TeamColumnProps) => {
-	const JoinButton = () => (
-		<Button onClick={() => onJoin(teamId)} variant="secondary">
-			<LuPlus className="inline-block mr-2" />
-			Join
-		</Button>
-	)
-
+const Title = () => {
 	return (
-		<ListPanel
-			title={`Team ${teamId === TeamId.Team1 ? 1 : 2}`}
-			items={members}
-			footer={<JoinButton />}
-			emptyLabel="No players yet."
-		/>
+		<h1 className="text-2xl font-bold text-amber-100/90 text-center mb-3 tracking-wider">
+			Choose Your Team
+		</h1>
 	)
 }
 
@@ -34,32 +18,8 @@ export const TeamSelection = () => {
 	const client = useGameClient()
 
 	if (!gameState || !gameState.gameState) {
-		return <div>Invalid state...</div> // TODO
+		return <div>Invalid state...</div>
 	}
-
-	const Title = () => {
-		return (
-			<h1 className="text-2xl font-bold text-amber-100/90 text-center mb-3 tracking-wider">
-				Choose Your Team
-			</h1>
-		)
-	}
-
-	const RoomIdLabel = () => (
-		<div className="mb-4 flex items-center justify-center space-x-2">
-			<span className="text-sm font-bold text-amber-100/70 tracking-wider">
-				{`Room ID: ${gameState.gameState.roomId}`}
-			</span>
-			<LuCopy
-				className="inline-block cursor-pointer text-gray-400/30 hover:text-amber-100/70 transition-colors duration-200"
-				onClick={() =>
-					navigator.clipboard.writeText(gameState.gameState.roomId)
-				}
-				size={'1.25rem'}
-				title="Copy Room ID"
-			/>
-		</div>
-	)
 
 	const handleJoinTeam = (teamId: TeamId) => {
 		client.chooseTeam(teamId)
@@ -81,7 +41,7 @@ export const TeamSelection = () => {
 		<div className="w-full h-full grid place-items-center">
 			<Panel className="max-w-md w-full border border-yellow-600/30 rounded-2xl p-8">
 				<Title />
-				<RoomIdLabel />
+				<RoomIdLabel roomId={gameState.gameState.roomId} />
 				<div className="grid grid-cols-1 md:grid-cols-2 gap-6 w-full max-w-3xl mx-auto mb-6">
 					<TeamColumn
 						teamId={TeamId.Team1}
