+++ client/app/components/game/page/TeamSelection/index.ts
@@ -1,0 +1,1 @@
+export { TeamSelection as default } from './TeamSelection'
